+++ example/json_values.php
@@ -8,7 +8,7 @@ example = { json: true, is-it: [ good, great, awesome ] }
 EOF;
 
 $result = new \Pixel418\Iniliq\Parser;
-$result->parse( $content );
+$result = $result->parse( $content );
 
 include( __DIR__ . '/_output.php' );